refactor: restructure states
