+++ contracts/core/src/state/rebalance.rs
@@ -1,21 +1,8 @@
 use cosmwasm_schema::cw_serde;
 use cosmwasm_std::{Addr, Decimal, Uint128};
-use cw_storage_plus::{Item, Map};
 use ibcx_interface::types::SwapRoutes;
 
 use crate::error::ContractError;
-
-pub const LATEST_REBALANCE_ID_KEY: &str = "latest_rebalance_id";
-pub const LATEST_REBALANCE_ID: Item<u64> = Item::new(LATEST_REBALANCE_ID_KEY);
-
-pub const REBALANCES_PREFIX: &str = "rebalances";
-pub const REBALANCES: Map<u64, Rebalance> = Map::new(REBALANCES_PREFIX);
-
-pub const TRADE_INFOS_PREFIX: &str = "trade_infos";
-pub const TRADE_INFOS: Map<String, TradeInfo> = Map::new(TRADE_INFOS_PREFIX);
-
-pub const RESERVE_BUFFER_PREFIX: &str = "reserve_buffer";
-pub const RESERVE_BUFFER: Map<String, Uint128> = Map::new(RESERVE_BUFFER_PREFIX);
 
 #[cw_serde]
 pub struct Rebalance {
